+++ src/runtime/virtcontainers/agent.go
@@ -176,7 +176,7 @@ type agent interface {
 	markDead(ctx context.Context)
 
 	// cleanup removes all on disk information generated by the agent
-	cleanup(ctx context.Context, s *Sandbox)
+	cleanup(ctx context.Context)
 
 	// return data for saving
 	save() persistapi.AgentState
